merge '5.8-1arter97' from https://github.com/arter97/exfat-linux
   * Rebased to Linux mainline's drivers
   * Supports for kernels lower than v4.9 dropped

28 files changed, 6710 insertions(+), 11751 deletions(-)

This is now based on the latest Linux mainline's exFAT drivers by Samsung.

This new rebase contains a lot of new bug-fixes and improvements, and is radically different from the old base.
Symlink support is also removed to cope better with exFAT's standard and mainline's standard.

Support for kernel versions lower than v4.9 were dropped for easier maintenance. For people interested in exFAT support for said kernels, please use the old branch. It still works nicely and it's actively being shipped to production smartphones.

exfat-linux is planned to be maintained until Android devices with v5.7+ LTS kernel become more common.

For the time being, PPA packages are moved to a separate repository:
https://launchpad.net/~arter97/+archive/ubuntu/exfat-dev/

This will move to the main PPA repository within the next couple of weeks.

Signed-off-by: shashank1436 <9945shashank@gmail.com>